requirements.txt
pynput
psutil
pywin32
Pillow
pandas
